[DX] Extract ExecutionContext
