+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Entity/JobExecutionTest.php
@@ -5,7 +5,7 @@ namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Entity;
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Item\ExecutionContext;
+use Akeneo\Component\Batch\Item\ExecutionContext;
 use Akeneo\Component\Batch\Job\BatchStatus;
 use Akeneo\Component\Batch\Job\ExitStatus;
 use Doctrine\Common\Collections\ArrayCollection;
