+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Entity/StepExecutionTest.php
@@ -4,7 +4,7 @@ namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Entity;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Item\ExecutionContext;
+use Akeneo\Component\Batch\Item\ExecutionContext;
 use Akeneo\Component\Batch\Job\BatchStatus;
 use Akeneo\Component\Batch\Job\ExitStatus;
